requirements.txt
streamlit==1.45.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
requests==2.32.3
python-dotenv==1.0.1
tldextract==5.1.2
plotly==5.22.0
Pillow==10.3.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
pyzbar==0.1.9
joblib==1.4.2